change var types to portable type

+++ philo/src/time.c
@@ -6,7 +6,7 @@
 /*   By: pmolnar <pmolnar@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/06/23 22:56:41 by pmolnar       #+#    #+#                 */
-/*   Updated: 2022/07/27 11:25:40 by pmolnar       ########   odam.nl         */
+/*   Updated: 2022/08/07 13:20:17 by pmolnar       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,9 @@ void	precise_sleep(long duration)
 
 uint32_t	time_delta_msec(t_time t_old, t_time t_new)
 {
-	long	t_new_msec;
-	long	t_old_msec;
-	long	delta;
+	uint32_t	t_new_msec;
+	uint32_t	t_old_msec;
+	uint32_t	delta;
 
 	t_old_msec = t_old.tv_sec * 1000 + t_old.tv_usec / 1000;
 	t_new_msec = t_new.tv_sec * 1000 + t_new.tv_usec / 1000;
